+++ source/Application/Controller/AccountNewsletterController.php
@@ -8,7 +8,6 @@
 namespace OxidEsales\EshopCommunity\Application\Controller;
 
 use OxidEsales\Eshop\Core\Registry;
-use oxRegistry;
 
 /**
  * Current user newsletter manager.
@@ -24,14 +23,6 @@ class AccountNewsletterController extends \OxidEsales\Eshop\Application\Controll
      * @var string
      */
     protected $_sThisTemplate = 'page/account/newsletter.tpl';
-
-    /**
-     * @deprecated Property is not used and will be removed.
-     * Whether the newsletter option had been changed.
-     *
-     * @var bool
-     */
-    protected $_blNewsletter = null;
 
     /**
      * Whether the newsletter option had been changed give some affirmation.