+++ source/Application/views/admin/en/lang.php
@@ -769,25 +769,9 @@ $aLang = [
     'NAVIGATION_SYSREQ_MESSAGE_INACTIVE'             => "The permanent system health check has been deactivated. You can re-activate it in the Perform. tab at Master settings -> Core settings. To check it anyway, please click on this link: ",
     'NAVIGATION_SHOPFRONT'                           => "Shop's start page",
 
-    // @deprecated Functionality for Newsletter management will be removed.
-    'NEWSLETTER_DONE_NEWSSEND'                           => 'Your Newsletter has been sent.',
-    'NEWSLETTER_DONE_GOTONEWSLETTER'                     => 'go to Newsletter ',
-    'NEWSLETTER_DONE_TITLE'                              => 'OXID Newsletter',
-    // END deprecated
     'NEWSLETTER_LIST_MENUITEM'                           => 'Customer Information',
     'NEWSLETTER_LIST_MENUSUBITEM'                        => 'Newsletter',
-    // @deprecated Functionality for Newsletter management will be removed.
-    'NEWSLETTER_MAIN_MODEL'                              => 'Sample',
-    'NEWSLETTER_PLAIN_TEXT'                              => 'Plain Text Sample',
-    'NEWSLETTER_PREVIEW_PLAINTEXT'                       => 'PLAIN TEXT',
-    'NEWSLETTER_PREVIEW_HTML'                            => 'HTML',
-    'NEWSLETTER_SELECTION_USEDGROUP'                     => 'Used User Groups: ',
-    'NEWSLETTER_SELECTION_SELMAILRESAVER'                => 'Selected e-mail recipients',
-    'NEWSLETTER_SELECTION_SENDNEWS'                      => 'Send Newsletter',
-    'NEWSLETTER_SEND_TITLE'                              => 'OXID Newsletter',
-    'NEWSLETTER_SEND_SEND1'                              => 'Send',
-    'NEWSLETTER_SEND_SEND2'                              => 'of overall',
-    // END deprecated
+
     'ORDER_ARTICLE_YOUWANTTODELETE'                      => 'Do you really want to delete this product?',
     'ORDER_ARTICLE_YOUWANTTOSTORNO'                      => 'Do you really want to cancel this product?',
     'ORDER_ARTICLE_ENETTO'                               => 'Net',
@@ -1042,7 +1026,7 @@ $aLang = [
     'SHOP_CONFIG_DISABLEARTDUBLICATES'                        => 'When Products are copied: Do not set copied Product to active',
     'SHOP_CONFIG_EXPORTNUMBEROFLINES'                         => 'Number of Lines which are exported per Tick in CSV Export',
     'SHOP_CONFIG_MWSTSETTING'                                 => 'VAT Settings for XML export',
-    'SHOP_CONFIG_NUMBEROFEMAILSPERTICK'                       => 'Number of Newsletters sent per Tick',
+    'SHOP_CONFIG_NUMBEROFEMAILSPERTICK'                       => 'Number of emails sent per tick',
     'SHOP_CONFIG_UPDATEVIEWS'                                 => 'Database views needs to be updated!',
     'SHOP_CONFIG_ASSIGNDEFAULTCAT'                            => 'Assign default category',
     'SHOP_CONFIG_UNASSIGNDEFAULTCAT'                          => 'Unassign default category',
@@ -1285,9 +1269,6 @@ $aLang = [
     'TOOLTIPS_NEWUSERPAYMENT'                 => 'Create new Payment Method',
     'TOOLTIPS_NEWREMARK'                      => 'Add Note',
     'TOOLTIPS_NEWPAYMENT'                     => 'Create new Payment Method',
-    // @deprecated Functionality for Newsletter management will be removed.
-    'TOOLTIPS_NEWNEWSLETTER'                  => 'New Newsletter Sample',
-    // END deprecated
     'TOOLTIPS_NEWSHOP'                        => 'Create New Shop',
     'TOOLTIPS_NEWUSERGROUP'                   => 'Create New User Group',
     'TOOLTIPS_NEWCATEGORY'                    => 'Create New Category',
@@ -1529,13 +1510,7 @@ $aLang = [
     'mxferoles'                => 'Shop Roles',
     /* tabs */
     'tbclorder_iPayment'       => 'IPayment',
-    // newsletter_list',
-    // @deprecated Functionality for Newsletter management will be removed.
-    'tbclnewsletter_main'      => 'HTML',
-    'tbclnewsletter_plain'     => 'TEXT',
-    'tbclnewsletter_preview'   => 'Preview',
-    'tbclnewsletter_selection' => 'Selection',
-    // END deprecated
+    'tbclnewsletter_main'      => 'Main',
     'tbcl_main'                => 'Main',
     'tbcladminlinks_mall'      => 'Mall',
     'tbcldelivery_mall'        => 'Mall',
@@ -1820,9 +1795,6 @@ $aLang = [
     'DELIVERY_RDFA_FEDERALEXPRESS'             => 'FedEx',
     'DELIVERY_RDFA_UPS'                        => 'UPS',
     'SHOP_SYSTEM_SHOWREMEMBERME'               => 'Show "Remember me" option in login box',
-    // @deprecated Functionality for Newsletter management will be removed.
-    'NEWSLETTER_SUBJECT'                       => 'Subject',
-    // END deprecated
 
     'SHOP_PERF_SEO_CACHE'                        => 'Enable SEO cache',
     'SHOP_PERF_SYSREQ_CHECK'                     => 'Enable system health messages at the start page',