+++ source/Core/Autoload/UnifiedNameSpaceClassMap.php
@@ -964,47 +964,11 @@ return [
         'isInterface'      => false,
         'isDeprecated'     => false
     ],
-    'OxidEsales\Eshop\Application\Controller\Admin\NewsletterList'               => [
-        'editionClassName' => \OxidEsales\EshopCommunity\Application\Controller\Admin\NewsletterList::class,
-        'isAbstract'       => false,
-        'isInterface'      => false,
-        'isDeprecated'     => true,
-    ],
     'OxidEsales\Eshop\Application\Controller\Admin\NewsletterMain'               => [
         'editionClassName' => \OxidEsales\EshopCommunity\Application\Controller\Admin\NewsletterMain::class,
         'isAbstract'       => false,
         'isInterface'      => false,
         'isDeprecated'     => true
-    ],
-    'OxidEsales\Eshop\Application\Controller\Admin\NewsletterPlain'              => [
-        'editionClassName' => \OxidEsales\EshopCommunity\Application\Controller\Admin\NewsletterPlain::class,
-        'isAbstract'       => false,
-        'isInterface'      => false,
-        'isDeprecated'     => true,
-    ],
-    'OxidEsales\Eshop\Application\Controller\Admin\NewsletterPreview'            => [
-        'editionClassName' => \OxidEsales\EshopCommunity\Application\Controller\Admin\NewsletterPreview::class,
-        'isAbstract'       => false,
-        'isInterface'      => false,
-        'isDeprecated'     => true,
-    ],
-    'OxidEsales\Eshop\Application\Controller\Admin\NewsletterSelection'          => [
-        'editionClassName' => \OxidEsales\EshopCommunity\Application\Controller\Admin\NewsletterSelection::class,
-        'isAbstract'       => false,
-        'isInterface'      => false,
-        'isDeprecated'     => true,
-    ],
-    'OxidEsales\Eshop\Application\Controller\Admin\NewsletterSelectionAjax'      => [
-        'editionClassName' => \OxidEsales\EshopCommunity\Application\Controller\Admin\NewsletterSelectionAjax::class,
-        'isAbstract'       => false,
-        'isInterface'      => false,
-        'isDeprecated'     => true,
-    ],
-    'OxidEsales\Eshop\Application\Controller\Admin\NewsletterSend'               => [
-        'editionClassName' => \OxidEsales\EshopCommunity\Application\Controller\Admin\NewsletterSend::class,
-        'isAbstract'       => false,
-        'isInterface'      => false,
-        'isDeprecated'     => true,
     ],
     'OxidEsales\Eshop\Application\Controller\Admin\ObjectSeo'                    => [
         'editionClassName' => \OxidEsales\EshopCommunity\Application\Controller\Admin\ObjectSeo::class,
@@ -1881,12 +1845,6 @@ return [
         'isAbstract'       => false,
         'isInterface'      => false,
         'isDeprecated'     => false
-    ],
-    'OxidEsales\Eshop\Application\Model\Newsletter'                              => [
-        'editionClassName' => \OxidEsales\EshopCommunity\Application\Model\Newsletter::class,
-        'isAbstract'       => false,
-        'isInterface'      => false,
-        'isDeprecated'     => true,
     ],
     'OxidEsales\Eshop\Application\Model\Object2Category'                         => [
         'editionClassName' => \OxidEsales\EshopCommunity\Application\Model\Object2Category::class,